4.1.2 rename valid.js to valid-props.js.
