+++ src/components/movie-card/movie-card.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {moviesProp} from "../../utils/valid";
+import {moviesProp} from "../../utils/valid-props";
 import {Link} from 'react-router-dom';
 
 const MovieCard = (props) => {
