+++ src/components/movie-content-reviews/movie-content-reviews.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import dayjs from 'dayjs';
-import {commentsProp} from '../../utils/valid';
+import {commentsProp} from '../../utils/valid-props';
 import {COLUMN_QUANTITY, PRECISION_RATING} from '/src/const';
 
 export const MovieContentReviews = (props) => {
